+++ libft/GNL/get_next_line_utils.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line_utils.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: stdevis <stdevis@student.42.fr>            +#+  +:+       +#+        */
+/*   By: shokahn <shokahn@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/08 14:51:24 by stdevis           #+#    #+#             */
-/*   Updated: 2025/03/28 17:40:40 by stdevis          ###   ########.fr       */
+/*   Updated: 2025/05/29 15:04:26 by shokahn          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../headers/get_next_line.h"
 
-size_t	gnl_ft_strlen(char *s)
+int	gnl_ft_strlen(char *s)
 {
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return (0);
@@ -24,9 +24,9 @@ size_t	gnl_ft_strlen(char *s)
 	return (i);
 }
 
-ssize_t	gnl_ft_strchr(char *s, int c)
+int	gnl_ft_strchr(char *s, int c)
 {
-	ssize_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
@@ -40,10 +40,10 @@ ssize_t	gnl_ft_strchr(char *s, int c)
 
 char	*gnl_ft_strjoin(char *s1, char *s2)
 {
-	ssize_t	i;
-	ssize_t	j;
-	ssize_t	len1;
-	ssize_t	len2;
+	int	i;
+	int	j;
+	int	len1;
+	int	len2;
 	char	*str_join;
 
 	if (!s1 || !s2)
@@ -64,9 +64,9 @@ char	*gnl_ft_strjoin(char *s1, char *s2)
 	return (str_join);
 }
 
-char	*gnl_ft_substr(char *s, ssize_t start, size_t len)
+char	*gnl_ft_substr(char *s, int start, int len)
 {
-	size_t	i;
+	int	i;
 	char	*substr;
 
 	if (!s)
@@ -89,8 +89,8 @@ char	*gnl_ft_substr(char *s, ssize_t start, size_t len)
 char	*gnl_ft_strdup(char *s1)
 {
 	char	*dup;
-	size_t	i;
-	size_t	len_s1;
+	int	i;
+	int	len_s1;
 
 	len_s1 = gnl_ft_strlen(s1);
 	i = 0;
