+++ libft/headers/get_next_line.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: stdevis <stdevis@student.42.fr>            +#+  +:+       +#+        */
+/*   By: shokahn <shokahn@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/18 13:12:47 by stdevis           #+#    #+#             */
-/*   Updated: 2025/01/23 14:42:50 by stdevis          ###   ########.fr       */
+/*   Updated: 2025/05/29 15:03:56 by shokahn          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,10 +29,10 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-ssize_t	gnl_ft_strchr(char *s, int c);
-size_t	gnl_ft_strlen(char *s);
+int	gnl_ft_strchr(char *s, int c);
+int	gnl_ft_strlen(char *s);
 char	*gnl_ft_strjoin(char *s1, char *s2);
-char	*gnl_ft_substr(char *s, ssize_t start, size_t len);
+char	*gnl_ft_substr(char *s, int start, int len);
 char	*gnl_ft_strdup(char *s1);
 char	*create_rest(char *line);
 char	*read_the_line(int fd, char *rest);
